requirements-dev.txt
-r ./requirements.txt
coverage ~= 4.0
nose2 ~= 0.6.0
pycontracts ~= 1.7